moved todoContext, add redux, redux-thunk, fix bugs, refactoring
