+++ src/pages/Main/Main.tsx
@@ -1,6 +1,13 @@
 import React from 'react';
 import { ToDo } from './ToDo'
+import TodoProvider from './ToDo/context/provider/todoProvider'
 
-const Main = () => <ToDo/>
+const Main = () => {
+  return (
+    <TodoProvider>
+      <ToDo/>
+    </TodoProvider>
+  )
+}
 
 export default Main
