+++ src/redux/reducers/index.tsx
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux'
+import { panelsReducer } from './panels'
+
+const rootReducer = combineReducers({
+  panelsReducer
+})
+
+export default rootReducer
