+++ src/redux/state.ts
@@ -1,4 +1,4 @@
-import { IInitialState } from './types/interfaces/interfaces'
+import { IInitialState } from 'types/interfaces/interfaces'
 
 export const store: IInitialState = {
   todos: [],
