+++ src/types/enums/panels.ts
